+++ src/accounts/devol_indexed_account.rs
@@ -14,17 +14,17 @@ pub trait DevolIndexedAccount : DevolAccount {
     fn id_offset() -> usize { 40 }
 
     #[inline(always)]
-    fn check_id(account_info: &AccountInfo, id: usize) -> Result<(), DvlError> {
+    fn check_id(account_info: &AccountInfo, id: u32) -> Result<(), DvlError> {
         let tag = AccountTag::from_u8(Self::expected_tag()).unwrap();
         let read_id = unsafe { *(account_info.data.borrow().as_ptr().add(Self::id_offset()) as *const u32) };
-        if read_id != id as u32 {
+        if read_id != id {
             return Err(DvlError::new_with_account(tag, ContractError::InvalidAccountId));
         }
         Ok(())
     }
 
     #[inline(always)]
-    fn check_all(account_info: &AccountInfo, root_addr: &Pubkey, program_id: &Pubkey, id: usize) -> Result<(), DvlError> {
+    fn check_all(account_info: &AccountInfo, root_addr: &Pubkey, program_id: &Pubkey, id: u32) -> Result<(), DvlError> {
         Self::check_basic(account_info, root_addr, program_id)?;
         Self::check_id(account_info, id)?;
         Ok(())
@@ -35,7 +35,7 @@ pub trait DevolIndexedAccount : DevolAccount {
         account_info: &'a AccountInfo,
         root_addr: &Pubkey,
         program_id: &Pubkey,
-        id: usize,
+        id: u32,
     ) -> Result<&'a Self, DvlError>
         where
             Self: Sized,
@@ -51,7 +51,7 @@ pub trait DevolIndexedAccount : DevolAccount {
         account_info: &'a AccountInfo,
         root_addr: &Pubkey,
         program_id: &Pubkey,
-        id: usize,
+        id: u32,
     ) -> Result<&'a mut Self, DvlError>
         where
             Self: Sized,
@@ -67,7 +67,7 @@ pub trait DevolIndexedAccount : DevolAccount {
         account: &mut impl Account,
         root_addr: &Pubkey,
         program_id: &Pubkey,
-        id: usize,
+        id: u32,
     ) -> Result<Box<Self>, Box<dyn Error>>
         where
             Self: Sized + Copy
